Renamed CallbackWrapper to TopRuleWrapper, handle case where search returns return nothing
